pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pymongo>=4.11",
    "pyrogram>=2.0.106",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=20.0",
    "telegram>=0.0.1",
    "tgcrypto>=1.2.5",
]
